requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
databases[sqlite]
psycopg2-binary
python-dotenv
passlib[bcrypt]
pydantic
pytest
